+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.22.3] - 2018
+## [0.22.5] - 2018-04-10
+### Fixed:
+- Show personal access tokens upon creation
+- Show notifications for Linode community site
+
+## [0.22.4] - 2018-04-10
+### Added:
+- Early-access notice
+### Changed:
+- Updated placeholder components and copy
+
+## [0.22.3] - 2018-04-09
 ### Added:
 - API Token Management
 - OAuth Client Management